Concurrent Survey Process
https://trello.com/c/B0ui7cdr/203-3-concurrent-survey-process-step-3-android-sdk

This commit:
-Includes end_user_id, user_id and unique_link on requests
to responses and decline
-Creates unique_link algorithm to be a SHA256 hash of NPS token,
end_user_email, date and a random string
-Adds priority number to responses and declines
-Adds account_id on requests to responses and declines
-Adds tests for SHAUtil
-Adds Apache Commons Codec lib

So I spend a lot of time reading and researching about
what would be the best way to test the new features.
Android project structures makes it a little bit
complicated to create unit tests on some stuff.

There are still things in the SDK that need some tests
and that can be tested. I'll keep thinking on ways to
get this done but for now this was the best I could do.