Adds end_user_last_seen, documentation, code cleanup and bug fix
License was added to all files header. Unused imports and
variables were removed.

Resources and demo app were updated to use new Wootric logo.

endUserCreatedAt date is now being checked to warn the user
if the date is on the future.

NullPointerException which caused the SDK to crash was
fixed on WootricRemoteRequestTask.java

end_user_last_seen was added on requests to eligibility.

There was a bug which cause first_survey_delay=31 to
always be sent on requests to eligibility. This meant
that on Android, the API always considered this value
over the db. Also, setFirstSurveyDelay wasn't available
to change this value from the SDK, this was also added.