Check activity is not null
So Practo's crashes logs are this:
Fatal Exception: android.view.WindowManager$BadTokenException: Unable to add window -- token android.os.BinderProxy@769aebf is not valid; is your activity running?

and

Caused by java.lang.NullPointerException: Attempt to invoke virtual method 'android.content.pm.PackageManager android.app.Activity.getPackageManager()' on a null object reference

I checked a lot and searched over the internet
and this has to do with the context being destroyed
before the survey completes the process and shows
up. I couldn't replicate because this is very hard
to replicate.

Here's an explanation I found on StackOverflow:

This can occur when you are showing the dialog for
a context that no longer exists. A common case -
if the 'show dialog' operation is after an
asynchronous operation, and during that operation
the original activity (that is to be the parent of
your dialog) is destroyed.

http://stackoverflow.com/questions/9529504/unable-to-add-window-token-android-os-binderproxy-is-not-valid-is-your-activ