Response/decline throttle
Updates the SDK to accept resurvey_throttle and decline_resurvey_throttle.
In case the end_user responds to the survey, the value of resurvey_throttle
will be used to set the expiration date of the surveyedAt value of
SharedPreferences (cookie), resurvey_throttle default value is 90.
In case the end_user declines the survey, the value of decline_resurvey_throttle
will be used as the expiration date, default value is 30.

Other fixes:
-Catch NullPointerException for Activity that cause crash
-Update tests
-Add tests for PreferencesUtils
-Update version of Gradle
-Add log for last_seen and surveyed 'cookie' values
-Clean code using Constants.TAG

Trello: https://trello.com/c/zBEZzmCW/732-1-custom-throttle-for-decline-response-android-update-the-clients-need-the-new-response-resurvey-date-and-decline-resurvey-date-